file_converter/requirements.txt
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
Cython==3.0.11
distlib==0.3.9
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.18.0
Flask==2.3.3
flask-cors==5.0.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.2
mutagen==1.47.0
mysql-connector-python==9.2.0
newspaper3k==0.2.8
nltk==3.9.2
packaging==25.0
pillow==12.0.0
pipx==1.7.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-file==3.0.1
rsa==4.9
selectolax==0.3.30
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.31.2
websockets==15.0.1
Werkzeug==3.1.3
yt-dlp==2025.8.22